feat: Change QRCode Data Tables from instance to Lazy static, initialize once
